kvdb_rocksdb

Struct DatabaseConfig

pub struct DatabaseConfig {
    pub max_open_files: i32,
    pub memory_budget: Option<usize>,
    pub compaction: CompactionProfile,
    pub columns: u32,
    pub disable_wal: bool,
}

Database configuration

Fields
max_open_files: i32

Max number of open files.

memory_budget: Option<usize>

Memory budget (in MiB) used for setting block cache size, write buffer size.

compaction: CompactionProfile

Compaction profile

columns: u32

Set number of columns

disable_wal: bool

Disable WAL if set to true

Implementations

impl DatabaseConfig

pub fn with_columns(columns: u32) -> Self

Create new DatabaseConfig with default parameters and specified set of columns. Note that cache sizes must be explicitly set.

pub fn memory_budget(&self) -> usize

pub fn memory_budget_per_col(&self) -> usize

pub fn memory_budget_mb(&self) -> usize