Crate kvdb_rocksdb

Structs

CompactionProfile

Compaction profile for the database settings Note, that changing these parameters may trigger the compaction process of RocksDB on startup. https://github.com/facebook/rocksdb/wiki/Leveled-Compaction#level_compaction_dynamic_level_bytes-is-true

Database

Key-Value database.

DatabaseConfig

Database configuration

Constants

DB_DEFAULT_COLUMN_MEMORY_BUDGET_MB

The default column memory budget in MiB.

DB_DEFAULT_MEMORY_BUDGET_MB

The default memory budget in MiB.

Functions

rotational_from_df_output

Given output of df command return Linux rotational flag file path.